+++ xivo_confd/plugins/skill_rule/validator.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+# Copyright 2018 The Wazo Authors  (see the AUTHORS file)
+# SPDX-License-Identifier: GPL-3.0+
+
+from xivo_confd.helpers.validator import (
+    ValidationGroup,
+)
+
+
+def build_validator():
+    return ValidationGroup()